+++ lib/pure/options.nim
@@ -68,6 +68,7 @@ supports pattern matching on `Option`s, with the `Some(<pattern>)` and
 ]##
 # xxx pending https://github.com/timotheecour/Nim/issues/376 use `runnableExamples` and `whichModule`
 
+{.deprecated: "Use `std/optionals`".}
 
 import typetraits